requirements.txt
websocket_client==0.56.0
paho-mqtt==1.5.1
